+++ src/n8n/nodes/10-read/format-help-message__83d12448-5f97-48f1-9ece-de61a9756db3.js
@@ -10,6 +10,8 @@ const msg = `Commands:
 /recipe <R<number>|query>
 /recipes <query>
 /recipe-save <structured_recipe_text>
+/recipe-link <public_id_1> <public_id_2>
+/recipe-note <public_id> <note>
 /last "query" [--days N] [--limit M]
 /find "needle" [--days N] [--limit M]
 /continue topic [--days N] [--limit M]